crewai_agents/requirements.txt
crewai>=0.5.0
pyyaml>=6.0
openai>=1.7.1
langchain>=0.1.0
anthropic>=0.18.0
langchain-anthropic>=0.1.0
